requirements.txt
cherrypy==18.6.0
flask==1.1.2
graphviz==2.38
keras==2.3.1
matplotlib==3.1.3
numpy==1.18.1
pandas==1.0.3
pillow==7.1.2
pydot==1.4.1
python==3.7.7
pyyaml==5.3.1
scikit-learn==0.22.1
seaborn==0.10.0
tensorflow==1.15.0
paste==3.2.2
